+++ ui-web-react/src/model/reducers/alertsQueue.reducer.ts
@@ -1,11 +1,7 @@
-import { AlertsQueue, Alert } from "../types/datatypes";
+import { AlertsQueue, Alert, INITIAL_ALERTS_QUEUE } from "../types/datatypes";
 import { AnyAction } from "redux";
 import { LoggingService } from "../utils/logging-service";
 import { ALERTS_SHOW, ALERTS_HIDE, AlertsAction } from "../actions/alerts.actions";
-
-export const INITIAL_ALERTS_QUEUE: AlertsQueue = {
-  alerts: []
-};
 
 export function alertsQueueReducer(state: AlertsQueue = INITIAL_ALERTS_QUEUE, action: AnyAction) {
   LoggingService.getInstance().logReducer(action, state);
